Warn if zlib compression is enabled before user enables compression in the plugin
git-svn-id: http://plugins.svn.wordpress.org/wp-super-cache/trunk@183028 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ wp-cache.php
@@ -168,7 +168,7 @@ jQuery(document).ready(function(){
 <?php
 	echo '<div class="wrap">';
 	echo "<h2><a href='?page=wpsupercache'>" . __( 'WP Super Cache Manager', 'wp-super-cache' ) . "</a></h2>\n";
-	if( 1 == ini_get( 'safe_mode' ) || "on" == strtolower( ini_get( 'safe_mode' ) ) ) {
+	if ( 1 == ini_get( 'safe_mode' ) || "on" == strtolower( ini_get( 'safe_mode' ) ) ) {
 		?><h3><?php _e( 'Warning! PHP Safe Mode Enabled!', 'wp-super-cache' ); ?></h3>
 		<p><?php _e( 'You may experience problems running this plugin because SAFE MODE is enabled.', 'wp-super-cache' );
 		if( !ini_get( 'safe_mode_gid' ) ) {
@@ -335,12 +335,16 @@ jQuery(document).ready(function(){
 			$cache_compression = 0;
 			wp_cache_replace_line('^ *\$cache_compression', "\$cache_compression = " . $cache_compression . ";", $wp_cache_config_file);
 		} elseif( isset( $_POST[ 'cache_compression' ] ) && $_POST[ 'cache_compression' ] != $cache_compression ) {
-			$cache_compression_changed = true;
-			$cache_compression = intval( $_POST[ 'cache_compression' ] );
-			wp_cache_replace_line('^ *\$cache_compression', "\$cache_compression = " . $cache_compression . ";", $wp_cache_config_file);
-			if( function_exists( 'prune_super_cache' ) )
-				prune_super_cache ($cache_path, true);
-			delete_option( 'super_cache_meta' );
+			if ( $_POST[ 'cache_compression' ] && 1 == ini_get( 'zlib.output_compression' ) || "on" == strtolower( ini_get( 'zlib.output_compression' ) ) ) { 
+				_e( "<strong>Warning!</strong> You attempted to enable compression but <code>zlib.output_compression</code> is enabled. See #21 in the Troubleshooting section of the readme file.", 'wp-super-cache' );
+			} else {
+				$cache_compression = intval( $_POST[ 'cache_compression' ] );
+				$cache_compression_changed = true;
+				wp_cache_replace_line('^ *\$cache_compression', "\$cache_compression = " . $cache_compression . ";", $wp_cache_config_file);
+				if( function_exists( 'prune_super_cache' ) )
+					prune_super_cache ($cache_path, true);
+				delete_option( 'super_cache_meta' );
+			}
 		}
 		if( isset( $_POST[ 'wp_cache_hide_donation' ] ) && $_POST[ 'wp_cache_hide_donation' ] != $wp_cache_hide_donation ) {
 			$wp_cache_hide_donation = intval( $_POST[ 'wp_cache_hide_donation' ] );
